+++ Java/DoublyLinkedList.java
