+++ README.md
@@ -49,15 +49,15 @@ Check contrbuting guidelines of [AASF-IIITM](https://github.com/AASF-IIITM/AASF-
     * [BubbleSort](https://github.com/AASF-IIITM/algorithms/blob/master/C%2B%2B/BubbleSort.cpp)
     * [SieveOfEratosthenes](https://github.com/AASF-IIITM/algorithms/blob/master/C%2B%2B/SieveofEratosthenes.cpp)
     * [MergeSort](https://github.com/AASF-IIITM/algorithms/blob/master/C%2B%2B/MergeSort.cpp)
-    * [CatalanNumbers/DyckPaths](https://github.com/AASF-IIITM/algorithms/blob/master/C%2B%2B/CatalanNumber:DyckPath.cpp)
-    * [kthSmallestElement](https://github.com/AASF-IIITM/algorithms/blob/master/C%2B%2B/FInding%20kth%20smallest%20element%20of%20array.cpp) 
-    * [InsertionSort](https://github.com/AASF-IIITM/algorithms/blob/master/C%2B%2B/Insertion%20Sort.cpp)
+    * [CatalanNumbers/DyckPaths](https://github.com/AASF-IIITM/algorithms/blob/master/C%2B%2B/CatalanNumberDyckPath.cpp)
+    * [kthSmallestElement](https://github.com/AASF-IIITM/algorithms/blob/master/C%2B%2B/FindingkthSmallestElementInArray.cpp) 
+    * [InsertionSort](https://github.com/AASF-IIITM/algorithms/blob/master/C%2B%2B/InsertionSort.cpp)
     * [JumpSearch](https://github.com/AASF-IIITM/algorithms/blob/master/C%2B%2B/JumpSearch.cpp)
-    * [NewtonsSquareRootAlgo](https://github.com/AASF-IIITM/algorithms/blob/master/C%2B%2B/Newton's%20square%20root%20algorithm.cpp)
+    * [NewtonsSquareRootAlgo](https://github.com/AASF-IIITM/algorithms/blob/master/C%2B%2B/NewtonSquareRootAlgorithm.cpp)
     * [ShellSort](https://github.com/AASF-IIITM/algorithms/blob/master/C%2B%2B/ShellSort.cpp)
     * [QuickSort2](https://github.com/AASF-IIITM/algorithms/blob/master/C%2B%2B/QuickSort2.cpp)
-    * [ExtraLongFactorial](https://github.com/AASF-IIITM/algorithms/blob/master/C%2B%2B/Extra%20Long%20Factorial.cpp)
-    * [KadaneAlgo](https://github.com/AASF-IIITM/algorithms/blob/newalgo/C%2B%2B/KadaneAlgo.cpp)
+    * [ExtraLongFactorial](https://github.com/AASF-IIITM/algorithms/blob/master/C%2B%2B/ExtraLongFactorial.cpp)
+    * [KadaneAlgo](https://github.com/AASF-IIITM/algorithms/blob/master/C%2B%2B/KadaneAlgo.cpp)
     * [Dijkstra](https://github.com/AASF-IIITM/algorithms/blob/master/C%2B%2B/Dijkstra.cpp)
     * [KruskalAlgo](https://github.com/AASF-IIITM/algorithms/blob/master/C%2B%2B/KruskalAlgo.cpp)
     * [Kosaraju](https://github.com/AASF-IIITM/algorithms/blob/master/C%2B%2B/Kosaraju.cpp)
@@ -68,7 +68,7 @@ Check contrbuting guidelines of [AASF-IIITM](https://github.com/AASF-IIITM/AASF-
     * [NextGreaterNumber](https://github.com/AASF-IIITM/algorithms/blob/master/C%2B%2B/NextGreaterNumber.cpp)
     * [PriorityQueueUsingLinkedList](https://github.com/AASF-IIITM/algorithms/blob/master/C%2B%2B/priorityQueueUsingLinkedList.cpp)
     * [DoublyLinkedList](https://github.com/AASF-IIITM/algorithms/blob/master/C%2B%2B/DoublyLinkedList.cpp)
-    * [armstrong no](https://github.com/AASF-IIITM/algorithms/blob/pull/C%2B%2B/arm.cpp)
+    * [Armstrong Number](https://github.com/AASF-IIITM/algorithms/blob/master/C%2B%2B/ArmstrongNumber.cpp)
     * [HillClimbing](https://github.com/AASF-IIITM/algorithms/blob/master/C%2B%2B/Hill%20climbing.cpp)
     * [DepthFirstSearch](https://github.com/AASF-IIITM/algorithms/blob/master/C%2B%2B/DepthFirstSearch.cpp)
     * [TowerOfHanoi](https://github.com/AASF-IIITM/algorithms/blob/master/C%2B%2B/TowerOfHanoi.cpp)
@@ -93,17 +93,17 @@ Check contrbuting guidelines of [AASF-IIITM](https://github.com/AASF-IIITM/AASF-
     * [InfixtoPostfix](https://github.com/AASF-IIITM/algorithms/blob/master/C%2B%2B/InfixToPostfix.cpp)
     * [BucketSort](https://github.com/AASF-IIITM/algorithms/blob/master/C%2B%2B/BucketSort.cpp)
     * [QuickhullAlgoforConvexHull](https://github.com/AASF-IIITM/algorithms/blob/master/C%2B%2B/ConvexHull.cpp)
-    * [DisjointSetUnion](https://github.com/AASF-IIITM/algorithms/blob/master/C%2B%2B/DSU.cpp)
+    * [DisjointSetUnion](https://github.com/AASF-IIITM/algorithms/blob/master/C%2B%2B/DisjointSetUnion.cpp)
     * [HamiltonianPath](https://github.com/cjindal013/algorithms/blob/work/C%2B%2B/HamiltonianPath.cpp)
-    * [LongestIncreasingSubsequence](https://github.com/pranjali-99/algorithms/blob/work-1/C%2B%2B/LongestIncreasingSubsequence.cpp)
+    * [LongestIncreasingSubsequence](https://github.com/AASF-IIITM/algorithms/blob/work/C%2B%2B/LongestIncreasingSubsequence.cpp)
     * [Combination](https://github.com/cjindal013/algorithms/blob/master/C%2B%2B/Combination.cpp)
 
 * Java
     * [BinarySearch](https://github.com/AASF-IIITM/algorithms/blob/master/Java/BinarySearch.java)
     * [BubbleSort](https://github.com/AASF-IIITM/algorithms/blob/master/Java/BubbleSort.java)
     * [SieveOfEratosthenes](https://github.com/AASF-IIITM/algorithms/blob/master/Java/SieveofEratosthenes.java)
-    * [SelectionSort](https://github.com/AASF-IIITM/algorithms/blob/pull/Java/selec.java)
-    * [DoublyLinkedList](https://github.com/AASF-IIITM/algorithms/blob/pull/Java/doub.java)
+    * [SelectionSort](https://github.com/AASF-IIITM/algorithms/blob/master/Java/SelectionSort.java)
+    * [DoublyLinkedList](https://github.com/AASF-IIITM/algorithms/blob/master/Java/DoublyLinkedList.java)
     * [HeapSort](https://github.com/AASF-IIITM/algorithms/blob/master/Java/HeapSort.java)
     * [TopologicalSort](https://github.com/AASF-IIITM/algorithms/blob/master/Java/TopologicalSort.java)
     * [InsertionSort](https://github.com/AASF-IIITM/algorithms/blob/master/Java/InsertionSort.java)
@@ -136,7 +136,7 @@ Check contrbuting guidelines of [AASF-IIITM](https://github.com/AASF-IIITM/AASF-
     * [BubbleSort](https://github.com/AASF-IIITM/algorithms/blob/master/Python3/BubbleSort.py)
     * [SelectionSort](https://github.com/AASF-IIITM/algorithms/blob/master/Python3/SelectionSort.py)
     * [SieveOfEratosthenes](https://github.com/AASF-IIITM/algorithms/blob/master/Python3/SieveOfEratosthenes.py)
-    * [FractionalKnapsack](https://github.com/AASF-IIITM/algorithms/blob/other-branch/Python3/FractionalKnapsack.py)
+    * [FractionalKnapsack](https://github.com/AASF-IIITM/algorithms/blob/master/Python3/FractionalKnapsack.py)
     * [DepthFirstTraversal](https://github.com/AASF-IIITM/algorithms/blob/master/Python3/depthFirstSearch.py)
     * [BreadthFirstTraversal](https://github.com/AASF-IIITM/algorithms/blob/master/Python3/breadthFirstSearch.py)
     * [SubsetBitManipulation](https://github.com/AASF-IIITM/algorithms/blob/master/Python3/subsetBitManipulation.py)
@@ -193,7 +193,7 @@ Check contrbuting guidelines of [AASF-IIITM](https://github.com/AASF-IIITM/AASF-
     * [QueueUsingLinkedList](https://github.com/AASF-IIITM/algorithms/blob/master/C/QueueUsingLinkedList.c)
     * [CircularQueue](https://github.com/AASF-IIITM/algorithms/blob/master/C/CircularQueue.c)
     * [ReverseQueue](https://github.com/AASF-IIITM/algorithms/blob/master/C/ReverseQueue.c)
-    * [StackUsingArray](https://github.com/AASF-IIITM/algorithms/blob/master/C/_Stack.cpp)
+    * [StackUsingArray](https://github.com/AASF-IIITM/algorithms/blob/master/C/StackUsingArray.cpp)
     * [InsertionSort](https://github.com/AASF-IIITM/algorithms/blob/master/C/insertionsort.cpp)
     * [CircularLinkedListwithAdjacentNodeInterchange](https://github.com/AASF-IIITM/algorithms/blob/master/C/CircularLinkedListAdjacentNodeInterchange.c)
     * [FCFS](https://github.com/AASF-IIITM/algorithms/blob/master/C/FirstComeFirstServed.c)
